Changed GameManager to humble
