+++ Assets/Scripts/GameManagerController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: c3f3a63ccc7e428409b5573b2868e97f
-timeCreated: 1496935266
+guid: 447074e70ddb24202b06bd6c592b37f6
+timeCreated: 1495914408
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
